requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
apptools=4.5.0=pypi_0
astor=0.8.1=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
bleach=3.1.1=pypi_0
ca-certificates=2019.11.28=hecc5488_0
cachetools=4.0.0=pypi_0
certifi=2019.11.28=py37hc8dfbb8_1
chardet=3.0.4=pypi_0
configobj=5.0.6=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
envisage=4.9.2=pypi_0
freetype=2.9.1=h8a8886c_1
gast=0.2.2=pypi_0
google-auth=1.11.3=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.27.2=pypi_0
h5py=2.10.0=pypi_0
idna=2.9=pypi_0
importlib-metadata=1.5.0=pypi_0
intel-openmp=2020.0=166
ipykernel=5.1.4=pypi_0
ipyparallel=6.2.4=pypi_0
ipython=7.13.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.16.0=pypi_0
jinja2=2.11.1=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter-client=6.0.0=pypi_0
jupyter-core=4.6.3=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.1.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
markdown=3.2.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.2.0=pypi_0
mayavi=4.7.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.4=pypi_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py37hfd86e86_0
notebook=6.0.3=pypi_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
open3d=0.9.0.0=pypi_0
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1e=h516909a_0
opt-einsum=3.2.0=pypi_0
pandas=1.0.2=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.6.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py37_1
prometheus-client=0.7.1=pypi_0
prompt-toolkit=3.0.4=pypi_0
protobuf=3.11.3=pypi_0
ptyprocess=0.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyface=6.1.2=pypi_0
pygments=2.6.1=pypi_0
pyparsing=2.4.6=pypi_0
pyqt5=5.14.1=pypi_0
pyqt5-sip=12.7.1=pypi_0
pyrsistent=0.15.7=pypi_0
python=3.7.6=h0371630_2
python-dateutil=2.8.1=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2019.3=pypi_0
pyyaml=5.3=pypi_0
pyzmq=19.0.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scipy=1.4.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=45.2.0=py37_0
six=1.14.0=py37_0
sqlite=3.31.1=h7b6447c_0
tensorboard=2.1.1=pypi_0
tensorflow-estimator=2.1.0=pypi_0
tensorflow-gpu=2.1.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tk=8.6.8=hbc83047_0
torchsummary=1.5.1=pypi_0
torchvision=0.5.0=py37_cu101
tornado=6.0.4=pypi_0
tqdm=4.43.0=py_0
traitlets=4.3.3=pypi_0
traits=6.0.0=pypi_0
traitsui=6.1.3=pypi_0
urllib3=1.25.8=pypi_0
vtk=8.1.2=pypi_0
wcwidth=0.1.8=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.0=pypi_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.4=h14c3975_4
zipp=3.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
